data-ingestion/data-extraction-agent/requirements.txt
# ==============================================================================
# Requirements for data extraction agent
# ==============================================================================
# Prerequisites:
# ------------------------------------------------------------------------------
# Install pip if not already installed:
# sudo apt install python3-pip
#
# Create a virtual environment (for package isolation from system-wide packages):
# Install venv module if not already installed:
#
# sudo apt install python3.12-venv
# python3 -m venv .venv
# source .venv/bin/activate

# ==============================================================================
# Required packages: make install
# ==============================================================================
# Common packages
pydantic-settings==2.11.0

# Kafka Client
confluent-kafka==2.11.0
types-confluent_kafka==1.3.6

# OpenTelemetry
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-logging==0.59b0
opentelemetry-instrumentation-confluent-kafka==0.59b0
opentelemetry-semantic-conventions==0.59b0